backend/arena/requirements.txt
uvicorn
fastapi
python-dotenv
setuptools
pysolr
openpyxl
pyignite
pandas
python-Levenshtein
coverage
antlr4-python3-runtime